moved the events to AfterlifeEffectsListener.java
